refactor(auth): mouv totp and start home page
