+++ src/routes/+page.svelte
@@ -1,24 +1,8 @@
 <script lang="ts">
-	import { useSession, signOut } from '$lib/auth-client';
 	import { Button } from '$lib/components/ui/button';
-
-	const session = useSession();
 </script>
 
-<div class="flex flex-col gap-16">
-	<div>
-		{#if $session.data}
-			{#each Object.entries($session.data.user) as [key, value] (key)}
-				<p><strong>{key}:</strong> {value}</p>
-			{/each}
-			<Button
-				class="glass"
-				onclick={async () => {
-					await signOut();
-				}}>Logout</Button
-			>
-		{:else}
-			<p>Not logged in</p>
-		{/if}
-	</div>
+<div class="flex flex-col items-center justify-center gap-8">
+	<h1 class="text-4xl font-bold">Bienvenue sur tank game</h1>
+	<Button href="/game" class="h-auto px-12 py-6 text-2xl font-bold">PLAY</Button>
 </div>
